Some view fixes

+++ app/src/main/java/com/leefo/budgetapplication/model/Category.java
@@ -25,10 +25,10 @@ public class Category {
     private final boolean isIncome;
 
 
-    private final float budgetGoal;
+    private final int budgetGoal;
 
 
-    public Category(String name, String color, boolean isIncome, float budgetGoal )
+    public Category(String name, String color, boolean isIncome, int budgetGoal )
     {
         this.name = name;
         this.color = color;
